Derive package version from environment.
This prepares the integration with CI where the variable is set using
`git describe`.

Version the creation of the package. For dirty versions
(0.1-11-ga900718) setuptools will complain that "The version specified
('0.1-11-ga900718') is an invalid version, this may not work as
expected with newer versions of setuptools, pip, and PyPI. Please see
PEP 440 for more details." This should not be a problem since the
deployment is intended to be triggered for tags (0.2).